eq_iir: move hifi implementation to arch/xtensa
Moves HiFi implementation of EQ IIR processing to the
arch/xtensa directory, where it belongs. There is no point
in keeping arch dependent implementations in generic layer.
