+++ src/arch/xtensa/audio/eq_iir_hifi3.c
@@ -6,16 +6,16 @@
 //         Liam Girdwood <liam.r.girdwood@linux.intel.com>
 //         Keyon Jie <yang.jie@linux.intel.com>
 
-#include <stdint.h>
-#include <stddef.h>
-#include <errno.h>
+#include <sof/audio/eq_iir.h>
+
+#if CONFIG_IIR_ARCH
+
 #include <sof/audio/format.h>
-#include <sof/audio/eq_iir/iir.h>
 #include <user/eq.h>
-
-#if IIR_HIFI3
-
 #include <xtensa/tie/xt_hifi3.h>
+#include <errno.h>
+#include <stddef.h>
+#include <stdint.h>
 
 /*
  * Direct form II transposed second order filter block (biquad)
