+++ src/audio/eq_iir/eq_iir.c
@@ -8,8 +8,7 @@
 
 #include <sof/audio/component.h>
 #include <sof/audio/buffer.h>
-#include <sof/audio/eq_iir/eq_iir.h>
-#include <sof/audio/eq_iir/iir.h>
+#include <sof/audio/eq_iir.h>
 #include <sof/audio/format.h>
 #include <sof/audio/pipeline.h>
 #include <sof/common.h>
